Moved comparison code to algcomparisons and separated it out into different classes.
