+++ tetrad-lib/src/main/java/edu/cmu/tetrad/algcomparison/ComparisonAlgorithm.java
@@ -1,0 +1,17 @@
+package edu.cmu.tetrad.algcomparison;
+
+import edu.cmu.tetrad.data.DataSet;
+import edu.cmu.tetrad.graph.Graph;
+
+import java.util.Map;
+
+/**
+ * Created by jdramsey on 6/4/16.
+ */
+public interface ComparisonAlgorithm {
+    Graph search(DataSet dataSet, Map<String, Number> parameters);
+
+    String getName();
+
+    Graph getComparisonGraph(Graph dag);
+}
