+++ tetrad-lib/src/main/java/edu/cmu/tetrad/algcomparison/continuous/ContinuousFci.java
@@ -1,0 +1,31 @@
+package edu.cmu.tetrad.algcomparison.continuous;
+
+import edu.cmu.tetrad.algcomparison.ComparisonAlgorithm;
+import edu.cmu.tetrad.data.DataSet;
+import edu.cmu.tetrad.graph.Graph;
+import edu.cmu.tetrad.search.DagToPag;
+import edu.cmu.tetrad.search.Fci;
+import edu.cmu.tetrad.search.IndTestMixedLrt;
+import edu.cmu.tetrad.search.IndependenceTest;
+
+import java.util.Map;
+
+/**
+ * Created by jdramsey on 6/4/16.
+ */
+public class ContinuousFci implements ComparisonAlgorithm {
+    public Graph search(DataSet dataSet, Map<String, Number> parameters) {
+        IndependenceTest test = new IndTestMixedLrt(dataSet, parameters.get("alpha").doubleValue());
+        Fci pc = new Fci(test);
+        return pc.search();
+    }
+
+    public String getName() {
+        return "AJMixedFci";
+    }
+
+    @Override
+    public Graph getComparisonGraph(Graph dag) {
+        return new DagToPag(dag).convert();
+    }
+}
