+++ tetrad-lib/src/main/java/edu/cmu/tetrad/algcomparison/mixed/MixedSemFgs.java
@@ -1,0 +1,62 @@
+package edu.cmu.tetrad.algcomparison.mixed;
+
+import edu.cmu.tetrad.algcomparison.ComparisonAlgorithm;
+import edu.cmu.tetrad.data.CovarianceMatrixOnTheFly;
+import edu.cmu.tetrad.data.DataSet;
+import edu.cmu.tetrad.data.DataUtils;
+import edu.cmu.tetrad.graph.Edge;
+import edu.cmu.tetrad.graph.EdgeListGraph;
+import edu.cmu.tetrad.graph.Graph;
+import edu.cmu.tetrad.graph.Node;
+import edu.cmu.tetrad.search.Fgs;
+import edu.cmu.tetrad.search.SearchGraphUtils;
+import edu.cmu.tetrad.search.SemBicScore;
+
+import java.util.Map;
+
+/**
+ * Created by jdramsey on 6/4/16.
+ */
+public class MixedSemFgs implements ComparisonAlgorithm {
+    public Graph search(DataSet Dk, Map<String, Number> parameters) {
+        Dk = DataUtils.convertNumericalDiscreteToContinuous(Dk);
+        SemBicScore score = new SemBicScore(new CovarianceMatrixOnTheFly(Dk));
+        score.setPenaltyDiscount(parameters.get("penaltyDiscount").doubleValue());
+        Fgs fgs = new Fgs(score);
+        Graph p = fgs.search();
+        return convertBack(Dk, p);
+    }
+
+    public String getName() {
+        return "SemFgs";
+    }
+
+    private Graph convertBack(DataSet Dk, Graph p) {
+        Graph p2 = new EdgeListGraph(Dk.getVariables());
+
+        for (int i = 0; i < p.getNodes().size(); i++) {
+            for (int j = i + 1; j < p.getNodes().size(); j++) {
+                Node v1 = p.getNodes().get(i);
+                Node v2 = p.getNodes().get(j);
+
+                Edge e = p.getEdge(v1, v2);
+
+                if (e != null) {
+                    Node w1 = Dk.getVariable(e.getNode1().getName());
+                    Node w2 = Dk.getVariable(e.getNode2().getName());
+
+                    Edge e2 = new Edge(w1, w2, e.getEndpoint1(), e.getEndpoint2());
+
+                    p2.addEdge(e2);
+                }
+            }
+        }
+        return p2;
+    }
+
+    public Graph getComparisonGraph(Graph dag) {
+        return SearchGraphUtils.patternForDag(dag);
+    }
+}
+
+
